#241: added redshift objects & split catalog objects into several files
