+++ app/data/model/interface.py
@@ -1,0 +1,43 @@
+import abc
+import enum
+from typing import Any, Self
+
+
+class RawCatalog(enum.Enum):
+    """
+    Enum that represents the catalogs in their original form. They are stored in two
+    forms: one is the unaggregated collection of original data on layer 1 and the other is the
+    aggregated data on layer 2.
+    """
+
+    ICRS = "icrs"
+    DESIGNATION = "designation"
+    REDSHIFT = "redshift"
+
+
+class CatalogObject(abc.ABC):
+    """
+    Represents an object stored in a particular catalog.
+    """
+
+    @classmethod
+    @abc.abstractmethod
+    def aggregate(cls, objects: list[Self]) -> Self:
+        pass
+
+    @classmethod
+    @abc.abstractmethod
+    def layer2_keys(cls) -> list[str]:
+        pass
+
+    @abc.abstractmethod
+    def catalog(self) -> RawCatalog:
+        pass
+
+    @abc.abstractmethod
+    def layer1_data(self) -> dict[str, Any]:
+        pass
+
+    @abc.abstractmethod
+    def layer2_data(self) -> dict[str, Any]:
+        pass
