+++ app/data/model/layer1.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass
 
-from app.data.model import common
+from app.data.model import interface
 
 
 @dataclass
 class Layer1CatalogObject:
     pgc: int
     object_id: str
-    catalog_object: common.CatalogObject
+    catalog_object: interface.CatalogObject
