+++ app/domain/adminapi/table_transfer.py
@@ -27,13 +27,18 @@ class TableTransferManager:
         convs: list[converters.QuantityConverter] = [
             converters.NameConverter(),
             converters.ICRSConverter(),
+            converters.RedshiftConverter(),
         ]
+
+        valid_converters = []
 
         for conv in convs:
             try:
                 conv.parse_columns(meta.column_descriptions)
-            except converters.ConverterError as e:
-                raise errors.LogicalError(f"Unable to process table: {str(e)}") from e
+            except converters.ConverterError:
+                continue
+
+            valid_converters.append(conv)
 
         offset = 0
 
@@ -48,8 +53,12 @@ class TableTransferManager:
 
             for obj_data in data.data.to_dict(orient="records"):
                 objects: list[model.CatalogObject] = []
-                for conv in convs:
-                    obj = conv.apply(obj_data)
+                for conv in valid_converters:
+                    try:
+                        obj = conv.apply(obj_data)
+                    except converters.ConverterError:
+                        continue
+
                     objects.append(obj)
 
                 # TODO: cross-identification
