+++ app/domain/converters/__init__.py
@@ -2,11 +2,13 @@ from app.domain.converters.coordinate import ICRSConverter
 from app.domain.converters.errors import ConverterError, ConverterNoColumnError
 from app.domain.converters.interface import QuantityConverter
 from app.domain.converters.name import NameConverter
+from app.domain.converters.redshift import RedshiftConverter
 
 __all__ = [
     "QuantityConverter",
     "ICRSConverter",
     "NameConverter",
+    "RedshiftConverter",
     "ConverterError",
     "ConverterNoColumnError",
 ]
